requirements.txt
boto3==1.7.37
botocore==1.10.37
certifi==2018.4.16
chardet==3.0.4
click==6.7
docutils==0.14
elasticsearch==6.2.0
elasticsearch-curator==5.5.4
futures==3.2.0
idna==2.7
jmespath==0.9.3
python-dateutil==2.7.3
PyYAML==3.12
requests==2.19.0
requests-aws4auth==0.9
s3transfer==0.1.13
six==1.11.0
urllib3==1.22
voluptuous==0.11.1
